+++ hazelcast/src/main/java/com/hazelcast/concurrent/lock/client/LockRequest.java
@@ -32,7 +32,7 @@ public final class LockRequest extends AbstractLockRequest {
     public LockRequest() {
     }
 
-    public LockRequest(Data key, int threadId, long ttl, long timeout) {
+    public LockRequest(Data key, long threadId, long ttl, long timeout) {
         super(key, threadId, ttl, timeout);
     }
