+++ hazelcast/src/main/java/com/hazelcast/multimap/MultiMapContainer.java
@@ -68,7 +68,7 @@ public class MultiMapContainer {
         creationTime = Clock.currentTimeMillis();
     }
 
-    public boolean canAcquireLock(Data dataKey, String caller, int threadId) {
+    public boolean canAcquireLock(Data dataKey, String caller, long threadId) {
         return lockStore != null && lockStore.canAcquireLock(dataKey, caller, threadId);
     }
 
@@ -76,11 +76,11 @@ public class MultiMapContainer {
         return lockStore != null && lockStore.isLocked(dataKey);
     }
 
-    public boolean txnLock(Data key, String caller, int threadId, long ttl){
+    public boolean txnLock(Data key, String caller, long threadId, long ttl){
         return lockStore != null && lockStore.txnLock(key, caller, threadId, ttl);
     }
 
-    public boolean unlock(Data key, String caller, int threadId){
+    public boolean unlock(Data key, String caller, long threadId){
         return lockStore != null && lockStore.unlock(key, caller, threadId);
     }
 
@@ -88,7 +88,7 @@ public class MultiMapContainer {
         return lockStore != null && lockStore.forceUnlock(key);
     }
 
-    public boolean extendLock(Data key, String caller, int threadId, long ttl) {
+    public boolean extendLock(Data key, String caller, long threadId, long ttl) {
         return lockStore != null && lockStore.extendLeaseTime(key, caller, threadId, ttl);
     }
