+++ hazelcast/src/main/java/com/hazelcast/multimap/operations/client/RemoveRequest.java
@@ -37,12 +37,12 @@ public class RemoveRequest extends MultiMapKeyBasedRequest {
 
     Data value;
 
-    int threadId;
+    long threadId;
 
     public RemoveRequest() {
     }
 
-    public RemoveRequest(String name, Data key, Data value, int threadId) {
+    public RemoveRequest(String name, Data key, Data value, long threadId) {
         super(name, key);
         this.value = value;
         this.threadId = threadId;
@@ -57,14 +57,14 @@ public class RemoveRequest extends MultiMapKeyBasedRequest {
     }
 
     public void write(PortableWriter writer) throws IOException {
-        writer.writeInt("t",threadId);
+        writer.writeLong("t",threadId);
         super.write(writer);
         final ObjectDataOutput out = writer.getRawDataOutput();
         value.writeData(out);
     }
 
     public void read(PortableReader reader) throws IOException {
-        threadId = reader.readInt("t");
+        threadId = reader.readLong("t");
         super.read(reader);
         final ObjectDataInput in = reader.getRawDataInput();
         value = new Data();
